update creation of final VCF.
This creates a script that calls make on a jobid file to force
concatenation of the final VCF. This make command is held when it
is submitted and is released only when all of the chromosomes are
finally submitted. At that time, it waits 10 seconds and submits
a slurm script that takes into account all of the jobs representing
each window to finish.

This last step has always been a bit complicated. Basically, to
reduce the compute time on any one node, I split up the genome into
manageable chunks at which to genotype. This comes with the
disadvantage of having to wait until all of these chunks are finished
to combine them. Because there can be hundreds of chunks at a time, I
~wanted~ needed to avoid flooding the scheduler with so many jobs;
this is why chromosome jobs are used to submit jobs.

Yes it's confusing.
Is there a better way?
Probably.
Will you go into it?
No.