changed SORTORDER to coordinate; increase threads
I had initially changed SORTORDER to queryname because that is how
FixMateInformation was re-sorting it after I fed it coordinate-sorted
data. Unfortunately, this caused FixMateInformation to hang when
attempting to process this file. Changing it back to coordinate fixed
the issue and increasing the number of threads saves me ~15 minutes
per run.